replaceable transform: introduce 'Transform42' class
With this change, you can now instantiate a 'Transform' object,
and make a pointer to it. This way we can programmatically change
which transform implementation will be in effect.

Backwards compatibility is maintained, so if you call
Transform.ToNormalizedExample (which is what our code in the rest of the
program does), you'll still get the same thing you always did.

Right now there is only one transform implementation: Transform42.